autotest: move helicopter tests to own file
